requirements.txt
Flask
numpy
scikit-learn
scipy
joblib
kafka-python
python-dotenv
requests